Implementation of scala.Dynamic with multimethods.
Instead of rewriting x.meth(args) to x.applyDynamic(meth, args) or
x.applyDynamic(meth)(args), I am just rewriting it to x.applyDynamic(args).

Then it's up to the author of the applyDynamic method to have it take an
implicit method of type scala.Dynamic.MethodName which, if left to resolve
to the provided method scala.Dynamic.callingName, will take the value
"meth".

This is the basic test I've been using during development:

object Test extends App with Dynamic {
  def applyDynamic(b: Boolean)(implicit m: Dynamic.MethodName) = "Boolean: "+m+"("+b+")"
  def applyDynamic(i: Int)(implicit m: Dynamic.MethodName) = "Int: "+m+"("+i+")"

  println(this.anInt(1))
  println(this.aBoolean(false))
}

which should of course print out:

Int: anInt(1)
Boolean: aBoolean(false)

The definition of trait Dynamic doesn't change (still empty), but it now has a
companion Module.

---------------------

This still needs analysis and/or testing in the following areas:

 * Behaviour if trying to use implicit Dynamic.callingMethod in
   other contexts.

 * Dynamic annotations (they are not supported, but defining them may cause
   trouble).

 * Using named parameters for all or some of the (explicit or implicit)
   parameters in the dynamic call.

 * Eta expansion in dynamic calls [is it even possible?]

 * Something called "lifted" in Typers.scala -- no idea what it is