Reapply the filter to remove adding empty strings to XML buffers.
This was in fact unnecessary because NodeBuffer.&+ already filters them,
but was needed to guarantee equivalence to the master code (except for
scala processing instructions).

All tests pass.

In addition, this parser has been verified to be equivalent to the one
currently in master by running:

../../build/quick/bin/scalac -Xprint:parser `find pos neg run -name "*.scala"|xargs grep -l '<[a-zA-Z]'` >~/workspace/scala/parser-results.txt

and comparing the results before/after. They only differ in the tests
introduced with the new feature.